refactor: Update Login import in model/index.js
The Login import in the model/index.js file has been updated to use the correct file name 'Login' instead of 'Auth'. This change ensures that the correct module is imported and used in the application.
